LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-preflight-star-expected.txt

CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8800/fetch/api/resources/preflight.py?allow_methods=*&allow_headers=x-test&. Method SUPER is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8800/fetch/api/resources/preflight.py?allow_methods=*&allow_headers=*&. Method OK is not allowed by Access-Control-Allow-Methods.
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8800/fetch/api/resources/preflight.py?origin=http://localhost:8800&credentials&allow_methods=*&allow_headers=*&. Method OK is not allowed by Access-Control-Allow-Methods.

PASS CORS that succeeds with credentials: false; method: GET (allowed: get); header: X-Test,1 (allowed: x-test) 
FAIL CORS that succeeds with credentials: false; method: SUPER (allowed: *); header: X-Test,1 (allowed: x-test) promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL CORS that succeeds with credentials: false; method: OK (allowed: *); header: X-Test,1 (allowed: *) promise_test: Unhandled rejection with value: object "TypeError: Type error"
PASS CORS that fails with credentials: true; method: OK (allowed: *); header: X-Test,1 (allowed: *) 
FAIL CORS that fails with credentials: true; method: GET (allowed: *); header: undefined (allowed: ) assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL CORS that fails with credentials: true; method: GET (allowed: *); header: undefined (allowed: *) assert_unreached: Should have rejected: undefined Reached unreachable code
